+++ src/interfaces/IncomingMessageEvent.ts
@@ -1,5 +1,5 @@
-import { Message } from "@sourceblock-ug/sb-sl7";
-import { MessageResponseEvent } from "./MessageResponseEvent.js";
+import type { Message } from '@sourceblock-ug/sb-sl7';
+import type { MessageResponseEvent } from './MessageResponseEvent.js';
 
 export interface IncomingMessageEvent extends MessageResponseEvent {
   msg: string;